requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.3.0,<3.0.0
email-validator==2.1.0
requests==2.31.0
starlette==0.27.0
python-dotenv==1.0.1
dynaconf==3.2.4
